This now work 100% at a quarter wave. I don't know about when the lenthg is not lambda/4 but its 2:10 am and I'm going to bed. It is however working in these ways: 1) Computes the right Zodd and Zeven for lambda/4 lines of any coupling factor 2) Computes the right variation with frequency. For example, there is the output when one runs :
Note the coupling at 50 and 350 MHz agrees with data collected on
Puff of -28.42 dB @ 49.2986 MHz Small differences are to be expected
since I can only enter Zodd and Zeven to the nearest int in Puff.

sparrow /export/home/davek/atlc/src/non_gui % design_coupler 20 50 350
Please be patient - this will take a few minutes or so

For a 20.000 dB 50.000 Ohm coupler with a length of 0.3750 m,
you need to have an odd-mode impedance Zodd to be 45.227 Ohms and
an even mode impedance Zeven to be 55.277 Ohms

 20.000 dB down <-- ************************** ---> 50.000 Ohm termination

 Drive this port --> ************************** ---> 50.000 Ohm termination
		     <-------   0.3750 m ----->

Drive Port 1, coupler out of port 2 and terminate the other ports in Zo
Such a coupler will have the response indicated below.

frequency = 50.000 MHz coupling is -28.343 dB down on the main arm
frequency = 125.000 MHz coupling is -21.603 dB down on the main arm
frequency = 200.000 MHz coupling is -20.000 dB down on the main arm
frequency = 275.000 MHz coupling is -21.603 dB down on the main arm
frequency = 350.000 MHz coupling is -28.343 dB down on the main arm

You may force the length to be any value you want using the -L option - it does
not have to be 0.3750 metres long
You may try to find a coupler with these dimensions using the -d option

Currently the -d option is not that fast, as it uses a brain-dead algorithm
Hopefully one day the algorithm will be speeded up.